Extracts component
